+++ default.nix
@@ -10,7 +10,7 @@ pkgs ? import nixpkgs {}
 , blockfrost-tests ?
     (builtins.fetchGit {
       url = "ssh://git@github.com/blockfrost/blockfrost-tests.git";
-      rev = "d3f2ab53e14798045424e13cce03e7dca6ab4217";
+      rev = "80f4b30ac9af25677c04668e9f12e2b51dd21b3c";
       allRefs = true;
     })
 , system ? builtins.currentSystem
@@ -71,6 +71,7 @@ in
       # We have to increase memsize, otherwise we will get error:
       # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
       virtualisation.memorySize = 8192;
+      virtualisation.diskSize = 2048;
 
       services.blockfrost = {
         enable = true;
@@ -106,6 +107,7 @@ in
       # We have to increase memsize, otherwise we will get error:
       # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
       virtualisation.memorySize = 8192;
+      virtualisation.diskSize = 2048;
 
       services.blockfrost = {
         enable = true;
@@ -142,6 +144,7 @@ in
       # We have to increase memsize, otherwise we will get error:
       # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
       virtualisation.memorySize = 8192;
+      virtualisation.diskSize = 2048;
 
       services.blockfrost = {
         enable = true;
@@ -165,6 +168,43 @@ in
       machine.succeed("cp -r ${blockfrost-tests} /tmp/tests")
       machine.succeed(
           "cd /tmp/tests && NIX_PATH=nixpkgs=${nixpkgs} nix-shell --run 'SERVER_URL=http://localhost:3000/ SERVICE_NAME=ryo yarn test:preprod --run' >&2"
+      )
+    '';
+  };
+
+  blockfrost-backend-ryo-test-sanchonet = testing.makeTest rec {
+
+    name = "blockfrost-backend-ryo-test-sanchonet";
+
+    nodes.machine = {
+      imports = [ ./nixos-module.nix ];
+      # We have to increase memsize, otherwise we will get error:
+      # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
+      virtualisation.memorySize = 8192;
+      virtualisation.diskSize = 2048;
+
+      services.blockfrost = {
+        enable = true;
+        package = blockfrost-backend-ryo;
+        settings = {
+          dbSync = {
+            host = builtins.getEnv "DBSYNC_HOST_SANCHONET";
+            user = "csyncdb";
+            database = "csyncdb";
+          };
+          network = "sanchonet";
+          tokenRegistryUrl = builtins.getEnv "TOKEN_REGISTRY_URL_TESTNETS";
+        };
+      };
+    };
+
+    testScript = ''
+      start_all()
+      machine.wait_for_unit("blockfrost-backend-ryo.service")
+      machine.wait_for_open_port(3000)
+      machine.succeed("cp -r ${blockfrost-tests} /tmp/tests")
+      machine.succeed(
+          "cd /tmp/tests && NIX_PATH=nixpkgs=${nixpkgs} nix-shell --run 'SERVER_URL=http://localhost:3000/ SERVICE_NAME=ryo yarn test:sanchonet --run' >&2"
       )
     '';
   };