+++ src/types/responses/governance.ts
@@ -1,0 +1,14 @@
+import type { OpenApiResponseTypes } from '../openapi-wrapper';
+
+export type DReps = OpenApiResponseTypes['dreps'];
+export type DRepsDrepID = OpenApiResponseTypes['drep'];
+export type DRepsDrepIDMetadata = OpenApiResponseTypes['drep_metadata'];
+export type DRepsDrepIDUpdates = OpenApiResponseTypes['drep_updates'];
+export type DRepsDrepIDDelegators = OpenApiResponseTypes['drep_delegators'];
+export type DRepsDrepIDVotes = OpenApiResponseTypes['drep_votes'];
+export type Proposals = OpenApiResponseTypes['proposals'];
+export type ProposalsProposal = OpenApiResponseTypes['proposal'];
+export type ProposalsProposalVotes = OpenApiResponseTypes['proposal_votes'];
+export type ProposalsProposalParameters = OpenApiResponseTypes['proposal_parameters'];
+export type ProposalsProposalWithdrawals = OpenApiResponseTypes['proposal_withdrawals'];
+export type ProposalsProposalMetadata = OpenApiResponseTypes['proposal_metadata'];